+++ src/interfaces/wrapperLike.ts
